% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trunc_sets_1f.R
\name{compute_S_centroid_1f_gencov}
\alias{compute_S_centroid_1f_gencov}
\title{Computes the conditioning set S for centroid linkage hierarchical clustering,
when testing for a difference in means wrt a single feature. This is for the
dependent features case}
\usage{
compute_S_centroid_1f_gencov(X, hcl, K, k1, k2, feat, scaledSigRow)
}
\arguments{
\item{X}{the n x q data set}

\item{hcl}{hclust object obtained by clustering X}

\item{K}{number of clusters}

\item{k1}{the index of first cluster involved in the test}

\item{k2}{the index of second cluster involved in the test}

\item{feat}{the index of the feature involved in the test}

\item{scaledSigRow}{a vector containing the (feat)th row of covariance
matrix Sigma divided by Sigma_{feat, feat}}
}
\value{
Returns an "Intervals" object containing the conditioning set.
}
\description{
Computes the conditioning set S for centroid linkage hierarchical clustering,
when testing for a difference in means wrt a single feature. This is for the
dependent features case
}
\keyword{internal}
