% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trunc_sets.R, R/trunc_sets_1f.R
\name{compute_S_single_1f}
\alias{compute_S_single_1f}
\title{Computes the conditioning set S for single linkage hierarchical clustering,
when we're testing for a difference in means wrt a single feature}
\usage{
compute_S_single_1f(X, hcl, K, k1, k2, feat)

compute_S_single_1f(X, hcl, K, k1, k2, feat)
}
\arguments{
\item{X}{the n x q data set}

\item{hcl}{hclust object obtained by clustering X}

\item{K}{number of clusters}

\item{k1}{the index of first cluster involved in the test}

\item{k2}{the index of second cluster involved in the test}

\item{feat}{the index of the feature involved in the test}
}
\value{
Returns an "Intervals" object containing the conditioning set.

Returns an "Intervals" object containing the conditioning set.
}
\description{
Computes the conditioning set S for single linkage hierarchical clustering,
when we're testing for a difference in means wrt a single feature

Computes the conditioning set S for single linkage hierarchical clustering,
when we're testing for a difference in means wrt a single feature
}
\keyword{internal}
