% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_raw_data.R
\name{get_raw_data}
\alias{get_raw_data}
\title{Function to convert json objects to dataframe}
\usage{
get_raw_data(js)
}
\arguments{
\item{x}{single json line}
}
\description{
Function to convert json objects to dataframe
}
