% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trunc_sets_kmeans_1f.R
\name{kmeans_compute_S_1f_genCov}
\alias{kmeans_compute_S_1f_genCov}
\title{Compute set S for the general covariance case}
\usage{
kmeans_compute_S_1f_genCov(
  X,
  estimated_k_means,
  all_T_clusters,
  all_T_centroids,
  n,
  diff_means_feat,
  v_vec,
  v_norm,
  T_length,
  k,
  feat,
  scaledSigRow,
  scaledSigRow_2_norm
)
}
\arguments{
\item{v_norm, }{2-norm of vector v}

\item{k, }{cluster of interest}

\item{feat, }{integer, the index of the feature involved in the test}

\item{scaledSigRow}{a vector p by 1 containing the (feat)th row of covariance
matrix Sigma divided by Sigma_{feat, feat}}

\item{scaledSigRow_2_norm}{ell_2 norm of the vector scaledSigRow}

\item{XTv, }{vector p by 1}

\item{cl, }{factor vector n by 1 (most recent cluster assignment)}

\item{v, }{contrast vector n by 1}

\item{i, }{index of observation}
}
\value{
Returns an "Intervals" object containing the conditioning set.
}
\description{
Compute set S for the general covariance case
}
\keyword{internal}
