% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mongo_helpers.R
\name{mongo2list}
\alias{mongo2list}
\title{Function to extract mongo data to an R list}
\usage{
mongo2list(mong, ...)
}
\arguments{
\item{...}{additional arguments passed to mon$export()}

\item{mon}{Mongo connection}
}
\description{
Function to extract mongo data to an R list
}
