% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trunc_sets_kmeans_1f.R
\name{norm_phi_canonical_kmeans_1f}
\alias{norm_phi_canonical_kmeans_1f}
\title{Represent <x'(phi)_i, x'(phi)_j> as a quadratic function in phi ----}
\usage{
norm_phi_canonical_kmeans_1f(
  X,
  last_centroids,
  diff_means_feat,
  v_norm,
  cl,
  k,
  v,
  i,
  weighted_v_i,
  feat,
  scaledSigRow,
  scaledSigRow_2_norm
)
}
\arguments{
\item{v_norm, }{2-norm of vector v}

\item{cl, }{factor vector n by 1 (most recent cluster assignment)}

\item{k, }{cluster of interest}

\item{v, }{contrast vector n by 1}

\item{i, }{index of observation}

\item{feat, }{integer, the index of the feature involved in the test}

\item{scaledSigRow}{a vector p by 1 containing the (feat)th row of covariance
matrix Sigma divided by Sigma_{feat, feat}}

\item{scaledSigRow_2_norm}{ell_2 norm of the vector scaledSigRow}

\item{XTv, }{vector p by 1}
}
\value{
parameters: a, b, c the coefficients of the quadratic equation such that (ax^2 + bx + c <= 0)
}
\description{
Represent <x'(phi)_i, x'(phi)_j> as a quadratic function in phi ----
}
\keyword{internal}
