% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trunc_sets_kmeans_1f_less_cond.R
\name{norm_sq_phi_kmeans_1f_less_cond}
\alias{norm_sq_phi_kmeans_1f_less_cond}
\title{Represent ||x'(phi, psi)_i-x'(phim psi)_j||_2^2 as a quadratic function of phi
when we're testing for a difference in means wrt a single feature  ----}
\usage{
norm_sq_phi_kmeans_1f_less_cond(
  X,
  v,
  a,
  B,
  diff_means_feat,
  mean_feat,
  BXj,
  v_norm,
  a_norm,
  i,
  j,
  feat,
  scaledSigRow,
  scaledSigRow_2_norm,
  UPsi
)
}
\arguments{
\item{v, }{the vector}

\item{v_norm, }{ell_2 norm of vector v}

\item{i, }{first index}

\item{j, }{second index}

\item{feat, }{integer, the index of the feature involved in the test}

\item{scaledSigRow}{a vector p by 1 containing the (feat)th row of covariance
matrix Sigma divided by Sigma_{feat, feat}}

\item{scaledSigRow_2_norm}{ell_2 norm of the vector scaledSigRow}

\item{UPsi}{the matrix-vector product of U and Psi (defined in xx TODO:
YC add reference to paper)}

\item{XTv, }{vector p by 1}
}
\value{
parameters: a, b, c the coefficients of the quadratic equation
such that (ax^2 + bx + c <= 0)
}
\description{
Represent ||x'(phi, psi)_i-x'(phim psi)_j||_2^2 as a quadratic function of phi
when we're testing for a difference in means wrt a single feature  ----
}
\keyword{internal}
