% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trunc_dist.R
\name{TNNormalize}
\alias{TNNormalize}
\title{Compute normalization constant truncated normal distribution}
\usage{
TNNormalize(mean, sd, E, approx = FALSE)
}
\arguments{
\item{mean}{the mean parameter}

\item{sd}{the standard deviation}

\item{E}{the truncation set, an "Intervals" object or a matrix where rows represents
a union of disjoint intervals.}

\item{approx}{should the approximation algorithm be used? Default is \code{FALSE},
where the approximation is not used in the first place. But when the result is wacky,
the approximation will be used.}
}
\value{
This function returns the value of the survival function evaluated at quantile \code{q}.
}
\description{
This function returns the normalization constant of a truncated normal distribution
with a given set  \eqn{E}.
}
\details{
Let \eqn{X} be a normal random variable with \code{mean} and \code{sd}. Truncating
    \eqn{X} to the set \eqn{E} is equivalent to conditioning on \eqn{{X \in E}}.
This function computes \eqn{P(X \in E)}.
}
\references{
Bryc, Wlodzimierz. "A uniform approximation to the right normal tail integral."
    Applied mathematics and computation 127.2 (2002): 365-374.
}
\keyword{internal}
