% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kmeans_inference_1f.R
\name{kmeans_inference_1f}
\alias{kmeans_inference_1f}
\title{Test for a difference in means between clusters of observations
identified via k-means clustering.}
\usage{
kmeans_inference_1f(
  X,
  k,
  cluster_1,
  cluster_2,
  feat,
  iso = FALSE,
  sig = NULL,
  covMat = NULL,
  iter.max = 10,
  seed = 1234,
  tol_eps = 1e-06,
  verbose = TRUE
)
}
\arguments{
\item{X}{Numeric matrix; \eqn{n} by \eqn{q} matrix of observed data}

\item{k}{Integer; the number of clusters for k-means clustering}

\item{cluster_1, cluster_2}{Two different integers in {1,...,k}; two estimated clusters to test, as indexed by the results of
\code{kmeans_estimation}.}

\item{iso}{Boolean. If TRUE, an isotropic covariance matrix model is used.}

\item{sig}{Numeric; noise standard deviation for the observed data, a non-negative number;
relevant if \code{iso}=TRUE. If it's not given as input, a median-based estimator will be by default (see Section 4.2 of our manuscript).}

\item{covMat}{Numeric matrix; if \code{iso} is FALSE, *required* \eqn{q} by \eqn{q} matrix specifying \eqn{\Sigma}.}

\item{iter.max}{Positive integer;     the maximum number of iterations allowed in k-means clustering algorithm. Default to \code{10}.}

\item{seed}{Random seed for the initialization in k-means clustering algorithm.}

\item{tol_eps}{A small number specifying the convergence criterion for k-means clustering,
default to \code{1e-6}.}
}
\value{
Returns a list with the following elements:
\itemize{
\item \code{p_naive} the naive p-value which ignores the fact that the clusters under consideration
are estimated from the same data used for testing
\item \code{pval} the selective p-value \eqn{p_{selective}} in Chen and Witten (2022+)
\item \code{final_interval} the conditioning set of Chen and Witten (2022+), stored as the \code{Intervals} class
\item \code{test_stat} test statistic: the difference in the empirical means of two estimated clusters
\item \code{final_cluster} Estimated clusters via k-means clustering
}
}
\description{
This function tests the null hypothesis of no difference in means between
output by k-means clustering. The clusters are numbered as per the results of
the \code{kmeans_estimation} function in the \code{CADET} package.
}
\details{
For better rendering of the equations, visit https://yiqunchen.github.io/KmeansInference/reference/index.html.

Consider the generative model \eqn{X ~ MN(\mu,I_n,\sigma^2 I_q)}. First recall that k-means clustering
solves the following optimization problem
\deqn{ \sum_{k=1}^K \sum_{i \in C_k} \big\Vert x_i - \sum_{i \in C_k} x_i/|C_k| \big\Vert_2^2 , }
 where \eqn{C_1,..., C_K} forms a partition of the integers \eqn{1,..., n}, and can be regarded as
 the estimated clusters of the original observations. Lloyd's algorithm is an iterative apparoach to solve
 this optimization problem.
Now suppose we want to test whether the means of two estimated clusters \code{cluster_1} and \code{cluster_2}
are equal; or equivalently, the null hypothesis of the form \eqn{H_{0}:  \mu^T \nu = 0_q} versus
\eqn{H_{1}: \mu^T \nu \neq 0_q} for suitably chosen \eqn{\nu} and all-zero vector \eqn{0_q}.

This function computes the following p-value:
\deqn{P \Big( || X^T\nu || \ge || x^T\nu ||_2 \; | \;
  \bigcap_{t=1}^{T}\bigcap_{i=1}^{n} \{ c_i^{(t)}(X) =
 c_i^{(t)}( x ) \},  \Pi Y  =  \Pi y \Big),}
where \eqn{c_i^{(t)}} is the is the cluster to which the \eqn{i}th observation is assigned during the \eqn{t}th iteration of
Lloyd's algorithm, and \eqn{\Pi} is the orthogonal projection to the orthogonal complement of \eqn{\nu}.
In particular, the test based on this p-value controls the selective Type I error and has substantial power.
Readers can refer to the Sections 2 and 4 in Chen and Witten (2022+) for more details.
}
\references{
Lloyd, S. P. (1957, 1982). Least squares quantization in PCM. Technical Note, Bell Laboratories.
Published in 1982 in IEEE Transactions on Information Theory, 28, 128–137.
}
